backend/requirements.txt
fastapi
uvicorn
openai
google-genai
pydantic
python-multipart
httpx
requests
pillow
PyYAML
boto3
python-dotenv
